Update to support G35 (T2254)
The G35 is the same as the G35+ (T2270) Minus the automatic empty station. I Haven't added it to HAS_CONSUMABLES or HAS_MAP_FEATURE as I'm not sure if it supports them but it does support basic operation and has been working stably for me for around a month!